refactor: MovieBookingActivity 버전별 분기 함수로 재사용하도록 만들기
