+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ReservationActivity" />
+        <activity android:name=".MovieBookingActivity" />
         <activity android:name=".ReservationCompleteActivity" />
     </application>
