+++ app/src/main/res/layout/movie_booking.xml
@@ -42,7 +42,7 @@
         app:layout_constraintHeight_percent="0.4">
 
         <TextView
-            android:id="@+id/booked_movie_title_text"
+            android:id="@+id/movie_title"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:fontFamily="sans-serif"
@@ -60,29 +60,28 @@
 
 
         <TextView
-            android:id="@+id/booked_movie_running_day_text"
+            android:id="@+id/movie_screening_date"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="20sp"
-            android:fontFamily="sans-serif-medium"
+            android:fontFamily="sans-serif"
             android:gravity="start"
             android:letterSpacing="0"
             android:lineSpacingMultiplier="1.29"
-            android:text="2025.04.01"
             android:textColor="@color/black"
             android:textSize="16sp"
             android:textStyle="normal"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/booked_movie_title_text"
+            app:layout_constraintTop_toBottomOf="@id/movie_title"
             tools:text="상영일: 2025.04.01"/>
 
         <TextView
-            android:id="@+id/booked_movie_running_time_text"
+            android:id="@+id/movie_running_time"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:fontFamily="sans-serif-medium"
+            android:fontFamily="sans-serif"
             android:gravity="start"
             android:letterSpacing="0"
             android:lineSpacingMultiplier="1.29"
@@ -92,7 +91,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/booked_movie_running_day_text"
+            app:layout_constraintTop_toBottomOf="@id/movie_screening_date"
             tools:text="러닝타임: 125분"/>
 
         <Spinner
